backend-python/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1

# Twilio for SMS
twilio==9.0.0

# PDF Generation
reportlab==4.0.9

# Image Processing
Pillow==10.2.0

# Statistical Analysis
numpy==1.26.3
scipy==1.12.0

# AI/ML
openai==1.3.5
langchain==0.1.0
numpy==1.26.2
scipy==1.11.4

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Email & Notifications
sendgrid==6.11.0
twilio==8.10.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Background tasks
celery==5.3.4
flower==2.0.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0
